chore: rename 'functions' directory to 'api'
