+++ src/api/user/userService.ts
@@ -1,8 +1,8 @@
 import { randomUUID } from "crypto";
 
-import { dynamo } from "@/functions/common/db";
-import { requireEnv } from "@/functions/common/env";
-import { buildUpdateExpression } from "@/functions/common/update";
+import { dynamo } from "@/api/utils/db";
+import { requireEnv } from "@/api/utils/env";
+import { buildUpdateExpression } from "@/api/utils/update";
 import { CreateUserInput, User } from "@/types";
 
 const USER_TABLE = requireEnv("USER_TABLE");
